move ptxbundle to transaction file
