+++ taiga_halo2/src/transparent_ptx.rs
@@ -5,48 +5,9 @@ use crate::{
 use pasta_curves::pallas;
 
 #[derive(Debug, Clone)]
-pub struct TransparentPartialTxBundle {
-    partial_txs: Vec<TransparentPartialTransaction>,
-    authorization: Authorization,
-}
-
-#[derive(Debug, Clone)]
-pub struct TransparentResult {
-    pub nullifiers: Vec<TransparentNullifier>,
-    pub outputs: Vec<OutputResource>,
-}
-
-#[derive(Debug, Clone)]
-pub struct TransparentNullifier {}
-
-#[derive(Debug, Clone)]
-pub struct Authorization {}
-
-#[derive(Debug, Clone)]
 pub struct TransparentPartialTransaction {
     pub inputs: Vec<InputResource>,
     pub outputs: Vec<OutputResource>,
-}
-
-impl TransparentPartialTxBundle {
-    pub fn execute(&self) -> Result<TransparentResult, TransactionError> {
-        for partial_tx in self.partial_txs.iter() {
-            partial_tx.execute()?;
-        }
-
-        Ok(TransparentResult {
-            nullifiers: vec![],
-            outputs: vec![],
-        })
-    }
-
-    pub fn get_value_commitments(&self) -> Vec<ValueCommitment> {
-        unimplemented!()
-    }
-
-    pub fn digest(&self) -> [u8; 32] {
-        unimplemented!()
-    }
 }
 
 impl Executable for TransparentPartialTransaction {
